single gradle module
