+++ androidApp/src/main/java/com/example/fl0wer/androidApp/data/contacts/ReminderRepositoryImpl.kt
@@ -5,8 +5,8 @@ import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import com.example.fl0wer.androidApp.util.Const
-import com.example.fl0wer.domain.contacts.Contact
-import com.example.fl0wer.domain.contacts.ReminderRepository
+import com.example.fl0wer.androidApp.domain.contacts.Contact
+import com.example.fl0wer.androidApp.domain.contacts.ReminderRepository
 import com.example.fl0wer.androidApp.util.getAlarmManager
 import com.example.fl0wer.androidApp.util.setTimer
 import java.util.GregorianCalendar
