+++ androidApp/src/main/java/com/example/fl0wer/androidApp/di/DataModule.kt
@@ -7,19 +7,19 @@ import com.example.fl0wer.androidApp.data.core.network.GoogleApi
 import com.example.fl0wer.androidApp.data.directions.DirectionRepositoryImpl
 import com.example.fl0wer.androidApp.data.locations.LocationRepositoryImpl
 import com.example.fl0wer.androidApp.data.locations.database.LocationDao
-import com.example.fl0wer.domain.contacts.ContactsInteractor
-import com.example.fl0wer.domain.contacts.ContactsInteractorImpl
-import com.example.fl0wer.domain.contacts.ContactsRepository
-import com.example.fl0wer.domain.contacts.ReminderInteractor
-import com.example.fl0wer.domain.contacts.ReminderInteractorImpl
-import com.example.fl0wer.domain.contacts.ReminderRepository
-import com.example.fl0wer.domain.core.dispatchers.DispatchersProvider
-import com.example.fl0wer.domain.directions.DirectionInteractor
-import com.example.fl0wer.domain.directions.DirectionInteractorImpl
-import com.example.fl0wer.domain.directions.DirectionRepository
-import com.example.fl0wer.domain.locations.LocationInteractor
-import com.example.fl0wer.domain.locations.LocationInteractorImpl
-import com.example.fl0wer.domain.locations.LocationRepository
+import com.example.fl0wer.androidApp.domain.contacts.ContactsInteractor
+import com.example.fl0wer.androidApp.domain.contacts.ContactsInteractorImpl
+import com.example.fl0wer.androidApp.domain.contacts.ContactsRepository
+import com.example.fl0wer.androidApp.domain.contacts.ReminderInteractor
+import com.example.fl0wer.androidApp.domain.contacts.ReminderInteractorImpl
+import com.example.fl0wer.androidApp.domain.contacts.ReminderRepository
+import com.example.fl0wer.androidApp.domain.core.dispatchers.DispatchersProvider
+import com.example.fl0wer.androidApp.domain.directions.DirectionInteractor
+import com.example.fl0wer.androidApp.domain.directions.DirectionInteractorImpl
+import com.example.fl0wer.androidApp.domain.directions.DirectionRepository
+import com.example.fl0wer.androidApp.domain.locations.LocationInteractor
+import com.example.fl0wer.androidApp.domain.locations.LocationInteractorImpl
+import com.example.fl0wer.androidApp.domain.locations.LocationRepository
 import dagger.Module
 import dagger.Provides
 import java.util.GregorianCalendar
