+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/contacts/ContactsInteractor.kt
@@ -1,0 +1,8 @@
+package com.example.fl0wer.androidApp.domain.contacts
+
+interface ContactsInteractor {
+    suspend fun contacts(forceUpdate: Boolean = false): List<Contact>
+    suspend fun contact(lookupKey: String): Contact?
+    suspend fun search(nameFilter: String): List<Contact>
+    suspend fun contactsWithAddress(ignoreContactId: Int): List<Contact>
+}
