+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/contacts/ContactsRepository.kt
@@ -1,0 +1,7 @@
+package com.example.fl0wer.androidApp.domain.contacts
+
+interface ContactsRepository {
+    fun contacts(): List<Contact>
+    suspend fun contact(lookupKey: String): Contact?
+    suspend fun loadContacts(): List<Contact>
+}
