+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/contacts/ReminderRepository.kt
@@ -1,0 +1,9 @@
+package com.example.fl0wer.androidApp.domain.contacts
+
+import java.util.GregorianCalendar
+
+interface ReminderRepository {
+    fun birthdayReminder(contact: Contact): Boolean
+    fun addBirthdayReminder(contact: Contact, nextBirthday: GregorianCalendar)
+    fun removeBirthdayReminder(contact: Contact)
+}
