+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/core/dispatchers/DispatchersProvider.kt
@@ -1,0 +1,11 @@
+package com.example.fl0wer.androidApp.domain.core.dispatchers
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.MainCoroutineDispatcher
+
+interface DispatchersProvider {
+    val default: CoroutineDispatcher
+    val main: MainCoroutineDispatcher
+    val unconfined: CoroutineDispatcher
+    val io: CoroutineDispatcher
+}
