+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/core/dispatchers/DispatchersProviderImpl.kt
@@ -1,0 +1,10 @@
+package com.example.fl0wer.androidApp.domain.core.dispatchers
+
+import kotlinx.coroutines.Dispatchers
+
+object DispatchersProviderImpl : DispatchersProvider {
+    override val default = Dispatchers.Default
+    override val main = Dispatchers.Main.immediate
+    override val unconfined = Dispatchers.Unconfined
+    override val io = Dispatchers.IO
+}
