+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/locations/LatLon.kt
@@ -1,0 +1,6 @@
+package com.example.fl0wer.androidApp.domain.locations
+
+data class LatLon(
+    val latitude: Double,
+    val longitude: Double,
+)
