+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/locations/Location.kt
@@ -1,0 +1,8 @@
+package com.example.fl0wer.androidApp.domain.locations
+
+data class Location(
+    val id: Int,
+    val latitude: Double,
+    val longitude: Double,
+    val address: String,
+)
