+++ androidApp/src/main/java/com/example/fl0wer/androidApp/domain/locations/LocationInteractor.kt
@@ -1,0 +1,9 @@
+package com.example.fl0wer.androidApp.domain.locations
+
+import kotlinx.coroutines.flow.Flow
+
+interface LocationInteractor {
+    fun observeLocation(contactId: Int): Flow<Location?>
+    suspend fun mapClicked(contactId: Int, latitude: Double, longitude: Double)
+    suspend fun locations(): List<Location>
+}
