--- domain/src/main/kotlin/com/example/fl0wer/domain/contacts/ReminderInteractor.kt
@@ -1,8 +1,0 @@
-package com.example.fl0wer.domain.contacts
-
-interface ReminderInteractor {
-    suspend fun birthdayReminder(contact: Contact): Boolean
-    suspend fun addBirthdayReminder(contact: Contact)
-    suspend fun removeBirthdayReminder(contact: Contact)
-    suspend fun changeBirthdayReminder(contact: Contact)
-}
