--- domain/src/main/kotlin/com/example/fl0wer/domain/contacts/ReminderRepository.kt
@@ -1,9 +1,0 @@
-package com.example.fl0wer.domain.contacts
-
-import java.util.GregorianCalendar
-
-interface ReminderRepository {
-    fun birthdayReminder(contact: Contact): Boolean
-    fun addBirthdayReminder(contact: Contact, nextBirthday: GregorianCalendar)
-    fun removeBirthdayReminder(contact: Contact)
-}
