--- domain/src/main/kotlin/com/example/fl0wer/domain/core/Result.kt
@@ -1,6 +1,0 @@
-package com.example.fl0wer.domain.core
-
-sealed class Result<out T> {
-    data class Success<out T>(val value: T) : Result<T>()
-    data class Failure(val throwable: Throwable) : Result<Nothing>()
-}
