--- domain/src/main/kotlin/com/example/fl0wer/domain/core/dispatchers/DispatchersProvider.kt
@@ -1,11 +1,0 @@
-package com.example.fl0wer.domain.core.dispatchers
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.MainCoroutineDispatcher
-
-interface DispatchersProvider {
-    val default: CoroutineDispatcher
-    val main: MainCoroutineDispatcher
-    val unconfined: CoroutineDispatcher
-    val io: CoroutineDispatcher
-}
