--- domain/src/main/kotlin/com/example/fl0wer/domain/core/dispatchers/DispatchersProviderImpl.kt
@@ -1,10 +1,0 @@
-package com.example.fl0wer.domain.core.dispatchers
-
-import kotlinx.coroutines.Dispatchers
-
-object DispatchersProviderImpl : DispatchersProvider {
-    override val default = Dispatchers.Default
-    override val main = Dispatchers.Main.immediate
-    override val unconfined = Dispatchers.Unconfined
-    override val io = Dispatchers.IO
-}
