--- domain/src/main/kotlin/com/example/fl0wer/domain/directions/DirectionInteractor.kt
@@ -1,7 +1,0 @@
-package com.example.fl0wer.domain.directions
-
-import com.example.fl0wer.domain.core.Result
-
-interface DirectionInteractor {
-    suspend fun route(origin: String, destination: String): Result<Route>
-}
