--- domain/src/main/kotlin/com/example/fl0wer/domain/locations/LatLon.kt
@@ -1,6 +1,0 @@
-package com.example.fl0wer.domain.locations
-
-data class LatLon(
-    val latitude: Double,
-    val longitude: Double,
-)
