--- domain/src/main/kotlin/com/example/fl0wer/domain/locations/Location.kt
@@ -1,8 +1,0 @@
-package com.example.fl0wer.domain.locations
-
-data class Location(
-    val id: Int,
-    val latitude: Double,
-    val longitude: Double,
-    val address: String,
-)
